Comparison of narrow type with wide type in loop condition
CWE-190
CWE-197
CWE-835

Use a loop counter type at least as wide as limit.

Best fix with no behavior change: change count from uint16 to uint32 in
src/SharedFileList.cpp where it is declared before the loop.

No other logic changes are required.

Why this is best:

It directly addresses the mismatch (uint16 vs uint32).
It preserves semantics (values remain small due to existing cap).
It avoids casts in the loop condition and keeps intent clear.

+++ src/SharedFileList.cpp
@@ -724,7 +724,7 @@ void CSharedFileList::SendListToServer(){
 	// Files sent.
 	files.WriteUInt32(limit);
 
-	uint16 count = 0;
+	uint32 count = 0;
 	// Add to packet
 	std::vector<CKnownFile*>::iterator sorted_it = SortedList.begin();
 	for ( ; (sorted_it != SortedList.end()) && (count < limit); ++sorted_it ) {
